services/backend/requirements.txt
fastapi
uvicorn
sqlalchemy==1.4.46
cloud-sql-python-connector==1.0.0
pg8000
python-jose[cryptography]
httplib2
google-api-python-client
google-auth
google-auth-oauthlib
requests
google-auth-httplib2
itsdangerous

